refactor(ingest): extract wikitext stripper to wiki/strip.py
